customer-success-fte/pyproject.toml
[tool.poetry]
name = "customer-success-fte"
version = "1.0.0"
description = "24/7 AI Customer Success Employee — multi-channel support automation"
authors = ["AI Systems <ai@company.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# AI / LLM — uses OpenAI-compatible client pointed at Groq
openai = "^1.30.0"

# API framework
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-multipart = "^0.0.9"

# Database
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
alembic = "^1.13.1"
pgvector = "^0.3.0"
psycopg2-binary = "^2.9.9"

# Redis (replaces Kafka)
redis = {extras = ["hiredis"], version = "^5.0.4"}

# Google / Gmail
google-auth = "^2.29.0"
google-auth-oauthlib = "^1.2.0"
google-api-python-client = "^2.130.0"

# Twilio / WhatsApp
twilio = "^9.1.0"

# Utilities
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
tenacity = "^8.3.0"
structlog = "^24.1.0"
prometheus-client = "^0.20.0"
anyio = "^4.3.0"
websockets = "^12.0"

# Sentiment analysis
textblob = "^0.18.0"
nltk = "^3.8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
faker = "^25.0.0"
locust = "^2.27.0"
ruff = "^0.4.4"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing -v"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "W", "F", "I", "B", "C4", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
